backend/requirements.txt
google-generativeai==0.3.1
pgvector==0.2.4
langchain==0.1.0
fastapi==0.104.1
sqlalchemy==2.0.23
python-multipart==0.0.6
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
unstructured==0.10.30
pdf2image==1.16.3
tabulate==0.9.0
pandas==2.1.3
openpyxl==3.1.2
python-magic==0.4.27
PyPDF2==3.0.1
pytest==7.4.3
httpx==0.25.2
langchain-google-genai==0.0.6 